Moving into nonwear_utils
Moving some function into nonwear_utils module
